Split topic and service into their own executables
